Fix all gcc build errors and add armstrong numbers

+++ difference-of-squares/difference_of_squares.h
@@ -12,7 +12,7 @@ struct I {
     bool operator!=(I const& other) const { return value != other.value; }
 };
 
-inline int square_of_sum(int x) { return std::pow(std::accumulate(I{1}, I{x + 1}, 0), 2); }
+inline int square_of_sum(int x) { return static_cast<int>(std::pow(std::accumulate(I{1}, I{x + 1}, 0), 2)); }
 
 inline int sum_of_squares(int x) { return std::inner_product(I{1}, I{x + 1}, I{1}, 0); }
